make and use vandermonde matrix for polynomial; fix polynomial regression

+++ src/timeseries/autoregressive.rs
@@ -32,8 +32,8 @@ impl AR {
             .collect::<Vec<f64>>();
         let autocorrelations: Vec<f64> = (0..=self.p).map(|t| acf(&adjusted, t as i32)).collect();
         let r = &autocorrelations[1..];
-        let n = r.len() as i32;
-        let r_matrix = invert_matrix(&toeplitz(&autocorrelations[..n as usize]));
+        let n = r.len();
+        let r_matrix = invert_matrix(&toeplitz(&autocorrelations[..n]));
         let coeffs = matmul(&r_matrix, &r, n, n, false, false);
         self.coeffs = coeffs;
         self.coeffs.reverse();
